requirements.txt
streamlit
pandas
numpy
plotly
spotipy
google-api-python-client
google-auth-oauthlib
python-dotenv
